+++ isa-project/src/main/java/isaproject/model/cottage/Cottage.java
@@ -46,7 +46,7 @@ public class Cottage implements Serializable {
 	private Integer bedCount;
 	private Integer roomCount;
 	private String cottageRules;
-	private Integer pricePerHour;
+	private Double pricePerHour;
 	@OneToMany(mappedBy = "cottage", fetch = FetchType.EAGER)
 	@JsonManagedReference
 	private Set<CottageQuickReservation> cottageQuickReservation = new HashSet<>();
@@ -193,11 +193,11 @@ public class Cottage implements Serializable {
 		this.subscribers = subscribers;
 	}
 
-	public Integer getPricePerHour() {
+	public Double getPricePerHour() {
 		return pricePerHour;
 	}
 
-	public void setPricePerHour(Integer pricePerHour) {
+	public void setPricePerHour(Double pricePerHour) {
 		this.pricePerHour = pricePerHour;
 	}
